+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+# 3.2.4 (April 23, 2025)
+
+## Notes
+- Golang: **v1.22**
+
+### Bug Fixes
+
+[PR #323](https://github.com/zscaler/zscaler-sdk-go/pull/323) – Fixed ZPA resource `pracredentialpool` missing attribute `credentialMappingCount`
+[PR #323](https://github.com/zscaler/zscaler-sdk-go/pull/323) – Fixed ZPA resource `policysetcontrollerv2` attribute `credentialPool` pointer
+[PR #323](https://github.com/zscaler/zscaler-sdk-go/pull/323) – Fixed ZDX rate limit handling in the `getRetryAfter` function by correctly parsing and calculating retry delays based on the headers: `X-Ratelimit-Remaining-Second` and `X-Ratelimit-Limit-Second`.
+
 # 3.2.3 (April 22, 2025)
 
 ## Notes